Jitter solved.
The linespacing was interfering with the :hover. It wasn't being brought in because on line 84 I had put .twocols h3, h4. Therefore it was interfering with all <h4>s on the page. Since it was a subtle change I didn't notice. I have now made everything consistent and will remove any unneeded code in the next commit.